Revert "efl_ui_layout: set needs_size_calc flag at constructor"
This reverts commit 98b19244329725c76ba601ea66ad5161841a68b9.

This totally broke elm widget min size calcs, e dialogs... sizing is
totally broken and unusable as a result. this is a major break and
needs immediate reverting back to a working state. Sorry Woki. You
need to test this... elementary_Test out of the box is totally broken
as all hell not to mention a lot more.