eio: fix poll backend
the problem here is that first the backedn is only NULLed out when we
are not running with a work thread. Which is ... super stupid, as this
pointer is going to freed anyways.

After having this fixed, we are running into the next issue, the path
gets deleted, which means that when we are in the iterator, the next
call is going to access already freed memory, the solution for this is
to copy the path for the lifetime of the thread.

Additionally, there have been cases where the delete flag was set
earlier but then read after calling some more functions. The functions
here are not atomic, whenever we want to check if something is deleted,
we need to check the flag DIRECTLY otherwise we are just getting the old
value, which might have changed.

This fixes the crash, and replaces it with another hung in our beloved
bugfree testcase "Efl Io Model Monitor".

Differential Revision: https://phab.enlightenment.org/D9624