efl_io_model: remove custom counting logic
this logic brings a few problems:

- When files are annonced here, without the monitor having the context
of the files, immidiat actions like deletion of the file are resulting
in no deletion event. Which is definitly a bug, so we need to wait here
for the monitor.

- When a new count is annonced before the files are there from the
monitor, we are heading towards the same issue - additionally, we might
already have files available, due to the custom counting that are not
annoncned through the monitor, which means, at a later pointer we are
annoncing a new file which is already available via the API.

Right now i do not see a way to continue with this code, it caused major
issues with our CI and the POLL backend, for now things do work again.

Differential Revision: https://phab.enlightenment.org/D9625